LPS-77456 Move WikiUtil to wiki-web.
