+++ modules/apps/collaboration/wiki/wiki-web/src/main/java/com/liferay/wiki/web/internal/display/context/logic/WikiPortletInstanceSettingsHelper.java
@@ -21,9 +21,9 @@ import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.wiki.model.WikiNode;
 import com.liferay.wiki.model.WikiNodeModel;
 import com.liferay.wiki.service.WikiNodeServiceUtil;
-import com.liferay.wiki.util.WikiUtil;
 import com.liferay.wiki.web.configuration.WikiPortletInstanceConfiguration;
 import com.liferay.wiki.web.internal.display.context.util.WikiRequestHelper;
+import com.liferay.wiki.web.internal.util.WikiUtil;
 
 import java.util.List;
